ext: implement key usage
This commit lifts the key usage extension into the `ext` module.
